Re-factored admin files into admin directories
this commit splits the admin files in common and stocking apps into
individual files so that they are easier to maintain.  All checks and
tests continue to pass.
