+++ fsdviz/common/admin/Agency.py
@@ -1,0 +1,13 @@
+from django.contrib.gis import admin
+
+from .utils import fill_color_widget
+from ..models import Agency
+
+
+@admin.register(Agency)
+class AgencyModelAdmin(admin.ModelAdmin):
+    list_display = ("agency_name", "abbrev", "fill_color")
+    search_fields = ("agency_name", "abbrev")
+
+    def fill_color(self, obj):
+        return fill_color_widget(obj.color)
