+++ fsdviz/common/admin/Grid10.py
@@ -1,0 +1,12 @@
+from django.contrib.gis import admin
+
+from ..models import Grid10
+
+admin.site.empty_value_display = "(None)"
+
+
+@admin.register(Grid10)
+class Grid10Admin(admin.GeoModelAdmin):
+    list_display = ("grid", "lake", "slug")
+    list_filter = ("lake",)
+    search_fields = ("grid",)
