+++ fsdviz/common/admin/Species.py
@@ -1,0 +1,17 @@
+from django.contrib.gis import admin
+from .utils import fill_color_widget
+from ..models import Species
+
+admin.site.empty_value_display = "(None)"
+
+
+@admin.register(Species)
+class SpeciesModelAdmin(admin.ModelAdmin):
+    list_display = ("abbrev", "common_name", "scientific_name", "active", "fill_color")
+    search_fields = ["abbrev", "common_name"]
+    list_filter = [
+        "active",
+    ]
+
+    def fill_color(self, obj):
+        return fill_color_widget(obj.color)
