+++ fsdviz/stocking/admin/LifeStage.py
@@ -1,0 +1,14 @@
+from django.contrib import admin
+
+from fsdviz.common.admin.utils import fill_color_widget
+from ..models import LifeStage
+
+admin.site.empty_value_display = "(None)"
+
+
+@admin.register(LifeStage)
+class LifeStageModelAdmin(admin.ModelAdmin):
+    list_display = ("abbrev", "description", "fill_color")
+
+    def fill_color(self, obj):
+        return fill_color_widget(obj.color)
